+++ src/components/data-model-table/index.tsx
@@ -1,30 +1,10 @@
 import './style.scss';
 import { CreateButton, List } from '@refinedev/antd';
 import { BaseRecord } from '@refinedev/core';
-import { Table, TableColumnsType, TableProps } from 'antd';
+import { Table } from 'antd';
 import { ColumnType } from 'antd/es/table/interface';
 import { AnyObject } from 'antd/es/_util/type';
-import { ResourceType } from '../../resource-type';
-
-export interface IDataModelTableProps<
-  T extends AnyObject,
-  U extends BaseRecord,
-> {
-  tableProps: TableProps<U>;
-  columns: TableColumnsType<T>;
-  hideCreateButton?: boolean;
-  text?: string;
-  buttonAction?: () => void;
-}
-
-// For any list views using DataModelTable
-export interface IDataModelListProps {
-  filters?: Record<string, any>;
-  hideCreateButton?: boolean;
-  hideActions?: boolean;
-  parentView?: ResourceType;
-  viewMode?: 'table' | 'map';
-}
+import { IDataModelTableProps } from '../../model/interfaces';
 
 export const DataModelTable = <T extends AnyObject, U extends BaseRecord>({
   tableProps,
